requirements.txt
boto3
pydantic
aws-cdk-lib==2.144.0
constructs>=10.0.0,<11.0.0
requests
pyyaml
python-dotenv
pynacl
langchain_core
langchain_aws
langchain_community
langchain_openai
pinecone
langchain_pinecone
bs4
lxml